refactor: separate the classes so that they are more organized and easier to understand
