+++ project-e-dukate-frontend/src/components/Metrics/PaymentMetrics/index.ts
@@ -1,1 +1,4 @@
-export * from './PaymentMetricsCharts';
+export * from './MetricCard';
+export * from './TotalIncomeChart';
+export * from './StatusCountsChart';
+export * from './PendingVsCompletedChart';
