+++ project-e-dukate-frontend/src/pages/Metrics/PaymentsMetrics/index.ts
@@ -1,0 +1,1 @@
+export { default as PaymentMetrics } from './PaymentMetrics';
