+++ project-e-dukate-frontend/src/utils/paymentConstants.ts
@@ -1,0 +1,7 @@
+import { FilterOption } from '@/types/filterOption';
+
+export const periodTypeOptions: FilterOption[] = [
+  { value: 'weekly', label: 'Semanal' },
+  { value: 'monthly', label: 'Mensual' },
+  { value: 'yearly', label: 'Anual' },
+];
